refactor: 957 - move course hero to hero widget
