--- src/widgets/hero-course/index.ts
@@ -1,1 +1,0 @@
-export { HeroCourse } from './ui/hero-course';
