--- src/widgets/hero-course/types.ts
@@ -1,10 +1,0 @@
-import { ReactNode } from 'react';
-import { StaticImageData } from 'next/image';
-
-export type HeroSectionData = {
-  variant: 'mentorship' | 'base';
-  heading: string;
-  subHeading?: ReactNode | string;
-  topHeading?: string[];
-  image?: StaticImageData;
-};
