+++ src/widgets/hero/helpers/get-course-status.test.ts
