+++ src/widgets/hero/helpers/get-course-status.ts
