+++ src/widgets/hero/helpers/has-day.test.ts
