+++ src/widgets/hero/helpers/has-day.ts
