+++ src/widgets/hero/index.ts
@@ -1,2 +1,3 @@
+export { CourseHero } from './ui/course/course-hero';
 export { Hero } from './ui/base/hero';
 export { MentorshipHero } from './ui/mentorship/mentorship-hero';
