+++ src/widgets/hero/ui/course/course-hero.module.scss
