+++ src/widgets/hero/ui/course/registration-link.tsx
